added Responses folder and separated customTick into a component
